revert outgoing wire-name conversion; keep request-side payload fix only
Scope the fix to what issue #812 is about: incoming payload key handling.
sendWebRtcProviderCommand now routes the payload through the existing
convertCommandDataToMatter (which camelizes keys), same as the generic
invoke path. The outgoing response/event serialization stays unchanged.
