+++ packages/matter-server/test/converter/ConvertersTest.ts
@@ -913,78 +913,6 @@ describe("Converters", () => {
         });
     });
 
-    describe("convertMatterToWebSocketNameBased - CHIP wire names (issue #812)", () => {
-        it("should emit webRtcSessionID/videoStreamID/audioStreamID for ProvideOfferResponse", () => {
-            // matter.js returns its own camelCase names (webRtcSessionId, ...). HA's camera
-            // integration reads the CHIP SDK wire names (response["webRtcSessionID"], ...) as
-            // generated in python_client/chip/clusters/cluster_defs, so responses must be
-            // converted back to the wire convention.
-            const webRtcProviderCluster = ClusterMap[1363]!;
-            const responseModel = webRtcProviderCluster.commands["provideoffer"]!.responseModel;
-
-            const matterValue = { webRtcSessionId: 42, videoStreamId: null, audioStreamId: 7 };
-            const result = convertMatterToWebSocketNameBased(
-                matterValue,
-                responseModel,
-                webRtcProviderCluster.model,
-            ) as Record<string, unknown>;
-
-            expect(result).to.deep.equal({ webRtcSessionID: 42, videoStreamID: null, audioStreamID: 7 });
-            expect(result).to.not.have.property("webRtcSessionId");
-        });
-
-        it("should emit wire names including deferredOffer for SolicitOfferResponse", () => {
-            const webRtcProviderCluster = ClusterMap[1363]!;
-            const responseModel = webRtcProviderCluster.commands["solicitoffer"]!.responseModel;
-
-            const matterValue = { webRtcSessionId: 7, deferredOffer: true, videoStreamId: null, audioStreamId: null };
-            const result = convertMatterToWebSocketNameBased(
-                matterValue,
-                responseModel,
-                webRtcProviderCluster.model,
-            ) as Record<string, unknown>;
-
-            expect(result).to.have.property("webRtcSessionID", 7);
-            expect(result).to.have.property("deferredOffer", true);
-            expect(result).to.have.property("videoStreamID", null);
-            expect(result).to.have.property("audioStreamID", null);
-        });
-
-        it("should emit videoStreamID/audioStreamID for CameraAvStreamManagement allocate responses", () => {
-            // Same acronym mismatch, hit via the generic device_command invoke path. HA's camera
-            // integration allocates a video/audio stream before calling ProvideOffer.
-            const avsmCluster = ClusterMap[0x551]!;
-
-            const videoResult = convertMatterToWebSocketNameBased(
-                { videoStreamId: 3 },
-                avsmCluster.commands["videostreamallocate"]!.responseModel,
-                avsmCluster.model,
-            ) as Record<string, unknown>;
-            expect(videoResult).to.deep.equal({ videoStreamID: 3 });
-
-            const audioResult = convertMatterToWebSocketNameBased(
-                { audioStreamId: 9 },
-                avsmCluster.commands["audiostreamallocate"]!.responseModel,
-                avsmCluster.model,
-            ) as Record<string, unknown>;
-            expect(audioResult).to.deep.equal({ audioStreamID: 9 });
-        });
-
-        it("should apply field name overrides (Groups AddGroupResponse keeps groupID)", () => {
-            // Groups cluster (4): AddGroupResponse has GroupID - wire name is groupID.
-            const groupsCluster = ClusterMap[4]!;
-            const responseModel = groupsCluster.commands["addgroup"]!.responseModel;
-
-            const result = convertMatterToWebSocketNameBased(
-                { status: 0, groupId: 17 },
-                responseModel,
-                groupsCluster.model,
-            ) as Record<string, unknown>;
-
-            expect(result).to.deep.equal({ status: 0, groupID: 17 });
-        });
-    });
-
     describe("convertMatterToWebSocketNameBased - named command responses (issue #70)", () => {
         it("should use camelCase names instead of numeric IDs for struct keys", () => {
             // OperationalCredentials cluster (62), UpdateFabricLabel response = NOCResponse
